Create config.py
on-behalf-of: @BI9BBL DEV@BI9BBL.COM
