+++ data/tasks.json
@@ -31,9 +31,9 @@
     "00861067070143638": [
       {
         "task": "Task1_GeneralStatusReport",
-        "CommandUUID": "ed68681687334ea8870fc8e736f7f9ac",
+        "CommandUUID": "f4e1ab96d0eb4147b7f0c9ce62c52822",
         "interval": 120,
-        "lastExecuted": 1754602888.6748734,
+        "lastExecuted": 1763426489.8608124,
         "oneTime": false,
         "consumed": false
       }
@@ -41,9 +41,9 @@
     "00865068030214960": [
       {
         "task": "Task1_GeneralStatusReport",
-        "CommandUUID": "a824e2452f854d50ae5dec8ea9f8ad53",
+        "CommandUUID": "dac6baa4c0d140769cfdad8b4321a2c5",
         "interval": 120,
-        "lastExecuted": 1754602944.6909769,
+        "lastExecuted": 1763426489.8608124,
         "oneTime": false,
         "consumed": false
       }
@@ -51,19 +51,19 @@
     "00865068030215157": [
       {
         "task": "Task1_GeneralStatusReport",
-        "CommandUUID": "f55e28824d3d4e198b2340d1a3442f94",
+        "CommandUUID": "8655225acd674552afcd8bf31a0e67fe",
         "interval": 120,
-        "lastExecuted": 1754602855.9693348,
+        "lastExecuted": 1763426489.8608124,
         "oneTime": false,
-        "consumed": true
+        "consumed": false
       }
     ],
     "00869102040011169": [
       {
         "task": "Task1_GeneralStatusReport",
-        "CommandUUID": "714b38b26f2345c6a4923cb7aa0a9a21",
+        "CommandUUID": "32eb6feeda1d4707a8ea7328cc969b82",
         "interval": 120,
-        "lastExecuted": 1754602900.6781797,
+        "lastExecuted": 1763426489.8608124,
         "oneTime": false,
         "consumed": false
       }
@@ -71,12 +71,13 @@
     "00860139040799590": [
       {
         "task": "Task1_GeneralStatusReport",
-        "CommandUUID": "feaad723768b41bab411168a495eecfb",
+        "CommandUUID": "21ce6647072342ad88bca6d59eef9311",
         "interval": 120,
-        "lastExecuted": 1754602891.6762767,
+        "lastExecuted": 1763426489.8608124,
         "oneTime": false,
         "consumed": false
       }
     ]
-  }
+  },
+  "version": 1
 }
